chore(frontend): separate all svg images into icons and img folders tm-229
